Add support for Atomic
We need to install rpm-build somewhere and the best place is a container.
Then we need to be able to run some commands on the host.
Then we need access to the container's files ...
... Oops, we need to mount its root somewhere!
Finally we must unlock the host.
The changes will be discarded on reboot. Epic win! That's exactly what we need!

Signed-off-by: Jakub Filak <jfilak@redhat.com>